Move backgrounds to base implementation
